Rewrite repo-bootstrapping steps.
This adapts to new APIs in daf_butler and takes advantage of the new
repository import/export functionality, instead of relying on direct
SQL dumps.

It also brings collection usage in line with what ConvertRepoTask
produces, so we should be able to use the same input collections
to launch pipelines regardless of how we built the initial repo.