include jointcal in external resources
We're importing the jointcal results, because running
jointcal is currently beyond Gen3, and it would want
full focal planes, which is too much for ci_hsc. Instead,
we'll just test the mechanics of applying the jointcal
results in makeWarp and forcedPhotCcd.

Generated after running the Gen2 ci_hsc, and then:
./bin/exportExternalData.py /tigress/pprice/ci_hsc_gen2/DATA resources/external.yaml