Skip filterlabel correction test if composite is disassembled
If we run CI with disassembled composites we know that by definition
the filter label will be written correctly. This makes it hard to
force it to fail by faking a mismatch between data ID and filter
label.

If in the future such a mismatch could occur we would have to
change the FilterLabel formatter to do the same correction
as is done in ExposureFormatter.