Force two WCS failures to test setting to None
This is accomplished by setting the processCcd config parameter:
calibrate.astrometry.maxMeanDistanceArcsec=0.025

NOTE: This is sensitive to the dataset and algorithms used in processing.
If there is a change in either, the threshold might need adjusting.  This
will likely result in a failure of this ci script as the failed astrometry
dataIds have to be explicitly identified so that they can be checked for
appropriate coord entries and skipped when checking the match datasets
(which do not get persisted on an astrometry fail).