Test maxPsfTraceRadiusDelta fail for coaddition
In order to trigger the deselection of a detector for inclusion in a
coadd, it must fail the config.makeWarp.select.maxPsfTraceRadiusDelta
threshold check.  None of the detecotrs in this repo currently fail
to pass with the current default of 0.7 pixels.  Overridding that
to 0.2 here forces a single failure, thus exercising this selection
criterion.  A unittest is also added to check that the failed dataId
is not listed in the coaddInputs() list of the coadd.