Restore forced-astrometry-failure tests.
This makes one detector fail to fit astrometry (by making a treshold
unrealistically small), so we can test what happens downstream.  An
extremely similar test was previously removed when we switched to
CalibrateImageTask, since that changed how the failure looked to
downstream tasks and (at the time) we hadn't yet updated them to
behave correctly.