Fix tests for mergedForcedSources
In DM-51687 the forced_src and forced_diff catalogs were removed
and mergedForcedSources was created in forcedPhotObjectDetector,
which combined forcedPhotCcd, forcedPhotDiffim, and
writeForcedSourceTable.

As a result the test was not incorporating the
_insufficient_template_coverage_failures data ids, which
were needed to make the number of mergedForcedSources
correct (I verified that it has the same number of datasets
as the version from before DM-51687).